Move linter rules classes into single module
