--- linkml/linter/rules/__init__.py
@@ -1,13 +1,0 @@
-from .no_empty_title import NoEmptyTitleRule
-from .no_xsd_int_type import NoXsdIntTypeRule
-from .permissible_values_format import PermissibleValuesFormatRule
-from .recommended import RecommendedRule
-from .tree_root_class import TreeRootClassRule
-
-__all__ = [
-    "NoEmptyTitleRule",
-    "NoXsdIntTypeRule",
-    "PermissibleValuesFormatRule",
-    "RecommendedRule",
-    "TreeRootClassRule",
-]
